fix(gb): verify HALT bug via VRAM LCD output reading (#1990)
* fix(gb): verify HALT bug via VRAM LCD output reading (#1982)

The halt_bug.gb ROM outputs test results to the LCD (via VRAM tile map),
not via the serial port. The existing test was checking serial output which
is always empty for this ROM.

- Add vram_tilemap_text() helper that decodes tile map 0 ($9800-$9BFF)
  from VRAM as ASCII text, covering the 20x18 visible tile area
- Add run_blargg_rom_lcd() helper that runs for the full cycle budget and
  then reads the VRAM output
- Update test_halt_bug to use the LCD output approach and remove #[ignore]
- Fix pre-existing clippy warnings in NES mapper files and event_loop.rs:
  unused MAPPER_NUMBER constants (#[allow(dead_code)]), identity_op,
  manual_range_contains, manual_range_patterns, doc_lazy_continuation,
  unnecessary_cast, unused function
- Fix pre-existing unused_mut warning in src/gb/ppu/mod.rs

The HALT bug implementation (sm83.rs lines ~750-757, ~185-192) was already
correct; test_halt_bug now passes without any CPU change.

Closes #1982

* review: add early-exit polling to run_blargg_rom_lcd; fix mapper208 comment

+++ src/nes/cartridge/unlicensed/mapper259.rs
@@ -344,7 +344,7 @@ mod tests {
         mapper.write_prg(0x6000, 0xFF); // only 0x0F should be stored
         // 0x0F → mode=1, bank=15, banked_lower = 14, page = 14*2 = 28
         // With 16 banks, page 28 wraps to 28 % 16 = 12
-        assert_eq!(mapper.read_prg(0x8000), 12 as u8);
+        assert_eq!(mapper.read_prg(0x8000), 12_u8);
     }
 
     // -------------------------------------------------------------------------
